Make basemq/tasks/FairMQ*Task header only
