+++ basemq/tasks/FairMQProcessorTask.h
@@ -1,16 +1,10 @@
 /********************************************************************************
- * Copyright (C) 2014-2022 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH  *
+ * Copyright (C) 2014-2023 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH  *
  *                                                                              *
  *              This software is distributed under the terms of the             *
  *              GNU Lesser General Public Licence (LGPL) version 3,             *
  *                  copied verbatim in the file "LICENSE"                       *
  ********************************************************************************/
-/**
- * FairMQProcessorTask.h
- *
- * @since Dec 6, 2012-12-06
- * @author: D. Klein, A. Rybalchenko
- */
 
 #ifndef FAIRMQPROCESSORTASK_H_
 #define FAIRMQPROCESSORTASK_H_
@@ -21,16 +15,16 @@
 class FairMQProcessorTask : public FairTask
 {
   public:
-    FairMQProcessorTask();
+    FairMQProcessorTask() = default;
     FairMQProcessorTask(const FairMQProcessorTask&) = delete;
     FairMQProcessorTask operator=(const FairMQProcessorTask&) = delete;
 
-    virtual ~FairMQProcessorTask();
+    ~FairMQProcessorTask() override = default;
 
-    virtual void Exec(Option_t* opt = "0");
+    virtual void Exec(Option_t* opt = "0") {}
 
-    void SetPayload(std::unique_ptr<fair::mq::Message>& msg);
-    void GetPayload(std::unique_ptr<fair::mq::Message>& msg);
+    void SetPayload(std::unique_ptr<fair::mq::Message>& msg) { fPayload = std::move(msg); }
+    void GetPayload(std::unique_ptr<fair::mq::Message>& msg) { msg = std::move(fPayload); }
 
   protected:
     std::unique_ptr<fair::mq::Message> fPayload;
